pyproject.toml
[tool.poetry]
name = "k-level-cooperation"
version = "0.1.0"
description = ""
authors = ["Tylores <tylor.slay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "*"
pandas = "*"
cvxpy = "*"
networkx = "*"
dataclasses = "*"
matplotlib = "*"
'OpenDSSDirect.py' = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
